Fix performance of `isochrone(::PARSECLibrary...)`
In old code, the isochrone interpolation and packing into `NamedTuple` did not take much penalty from type instability. In this new code, it doubled the runtime of the method, so I added a kernel function to do the interpolation to serve as a function barrier which restores proper performance. Putting type annotations on `y0` and `y1`, the intermediate interpolations, seems to have a similar effect, but I try to avoid type annotations when possible.