Initial `MISTTrack`, `MISTTrackSet`, `isochrone`
`isochrone` implementation is looking fine, mostly followed PARSEC example. Trying to provide a common API for `isochrone` including inputs and outputs so that it is easy to swap between the different libraries. Maybe need to take metallicity in [M/H] for PARSEC rather than Z? Might be difficult because of the ambiguity with how the metallicity scale is defined.